rename to error, to align with parameters package
