+++ packages/idempotency/src/index.ts
@@ -1,4 +1,4 @@
-export * from './Exceptions';
+export * from './errors';
 export * from './IdempotencyConfig';
 export * from './idempotentDecorator';
 export * from './makeFunctionIdempotent';
