+++ packages/idempotency/tests/unit/makeFunctionIdempotent.test.ts
@@ -16,7 +16,7 @@ import {
   IdempotencyInconsistentStateError,
   IdempotencyItemAlreadyExistsError,
   IdempotencyPersistenceLayerError,
-} from '../../src/Exceptions';
+} from '../../src/errors';
 import { IdempotencyConfig } from '../../src';
 import { Context } from 'aws-lambda';
