+++ packages/idempotency/tests/unit/persistence/DynamoDbPersistenceLayer.test.ts
@@ -7,7 +7,7 @@ import { DynamoDBPersistenceLayer } from '../../../src/persistence/DynamoDBPersi
 import {
   IdempotencyItemAlreadyExistsError,
   IdempotencyItemNotFoundError,
-} from '../../../src/Exceptions';
+} from '../../../src/errors';
 import { IdempotencyRecord } from '../../../src/persistence';
 import type { DynamoPersistenceOptions } from '../../../src/types';
 import { IdempotencyRecordStatus } from '../../../src/types';
